requirements.txt
django
djangorestframework
drf_yasg
djangorestframework-simplejwt